Fix bug in ConstantActionField
