+++ Assets/Scripts/Incidents/CriteriaEvaluators/ICriteriaEvaluator.cs
@@ -8,16 +8,4 @@ namespace Game.Incidents
         Type Type { get; }
         bool Evaluate(IIncidentContext context, string propertyName);
 	}
-
-	abstract public class ContextEvaluator<T> : ICriteriaEvaluator where T : IIncidentContext
-	{
-		public Type Type => typeof(T);
-
-		public bool Evaluate(IIncidentContext context, string propertyName)
-		{
-			contextField.CalculateField
-			var propertyValue = (T)context.GetType().GetProperty(propertyName).GetValue(context);
-			return 
-		}
-	}
 }
