+++ Assets/Scripts/Incidents/IncidentActions/ConstantActionField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25624f182c3541147a930cdeb0affa52
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
